backend/requirements.txt
praw
psycopg2-binary
pytrends
pandas
tweepy
requests
beautifulsoup4
lxml
python-dateutil
scikit-learn
numpy

# Embeddings and ML
sentence-transformers>=2.2.0
torch>=2.0.0

# Web Application Dependencies
fastapi
uvicorn[standard]
google-genai
python-dotenv
Pillow

# Image Generation via Hugging Face Inference API
# No heavy dependencies needed - using API calls instead
